corpus/requirements.txt
wheel
setuptools
twine
pandas
requests
urllib3
bs4
tqdm
webdav
minio
elasticsearch
PyYAML

werkzeug<1.0.0
pandas
jupyter
notebook
jupyter_contrib_nbextensions
matplotlib

autopep8
